# packages/shared-types/schemas/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiPage_Sensor_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Sensor"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "ApiPage[Sensor]",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LoginRequest": {
        "description": "User login request.",
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "LoginResponse": {
        "description": "Login response payload.",
        "properties": {
          "csrf_token": {
            "title": "Csrf Token",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserPublic"
          }
        },
        "required": [
          "user",
          "csrf_token"
        ],
        "title": "LoginResponse",
        "type": "object"
      },
      "RegisterRequest": {
        "description": "User registration request.",
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 256,
            "minLength": 12,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "RegisterRequest",
        "type": "object"
      },
      "Sensor": {
        "additionalProperties": false,
        "description": "Registered sensor identity and health metadata.",
        "properties": {
          "capabilities": {
            "items": {
              "$ref": "#/components/schemas/SensorCapability"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "id": {
            "maxLength": 96,
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "last_seen": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen"
          },
          "name": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "revoked": {
            "default": false,
            "title": "Revoked",
            "type": "boolean"
          },
          "tailnet_ip": {
            "maxLength": 64,
            "minLength": 3,
            "title": "Tailnet Ip",
            "type": "string"
          },
          "version": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "tailnet_ip",
          "version"
        ],
        "title": "Sensor",
        "type": "object"
      },
      "SensorCapability": {
        "description": "Capabilities advertised by a sensor.",
        "enum": [
          "passive_capture",
          "channel_control",
          "active_modules",
          "rogue_ap",
          "deauth",
          "evil_twin",
          "captive_portal",
          "mitm"
        ],
        "title": "SensorCapability",
        "type": "string"
      },
      "SensorRegisterRequest": {
        "description": "Sensor registration payload.",
        "properties": {
          "capabilities": {
            "items": {
              "$ref": "#/components/schemas/SensorCapability"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "id": {
            "maxLength": 96,
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "name": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "tailnet_ip": {
            "maxLength": 64,
            "minLength": 3,
            "title": "Tailnet Ip",
            "type": "string"
          },
          "version": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "tailnet_ip",
          "version"
        ],
        "title": "SensorRegisterRequest",
        "type": "object"
      },
      "SensorRegisterResponse": {
        "description": "Sensor registration response with one-time client material.",
        "properties": {
          "ca_certificate_pem": {
            "title": "Ca Certificate Pem",
            "type": "string"
          },
          "client_certificate_pem": {
            "title": "Client Certificate Pem",
            "type": "string"
          },
          "client_private_key_pem": {
            "title": "Client Private Key Pem",
            "type": "string"
          },
          "sensor": {
            "$ref": "#/components/schemas/Sensor"
          }
        },
        "required": [
          "sensor",
          "client_certificate_pem",
          "client_private_key_pem",
          "ca_certificate_pem"
        ],
        "title": "SensorRegisterResponse",
        "type": "object"
      },
...
      "UserPublic": {
        "additionalProperties": false,
        "description": "Safe user fields returned from API responses.",
        "properties": {
          "email": {
            "maxLength": 254,
            "minLength": 3,
            "title": "Email",
            "type": "string"
          },
          "id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          },
          "totp_enabled": {
            "default": false,
            "title": "Totp Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "email"
        ],
        "title": "UserPublic",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Cheeky Pony API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "description": "Authenticate a user and set secure cookies.\n\nArgs:\n    payload: Login payload.\n    response: FastAPI response.\n    store: Application store.\n    passwords: Password service.\n    csrf: CSRF service.\n    tokens: Token service.\n    settings: Runtime settings.\n\nReturns:\n    Public user and CSRF token.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/register": {
      "post": {
        "description": "Register the first admin or an admin-created user.\n\nArgs:\n    payload: Registration payload.\n    request: FastAPI request.\n    store: Application store.\n    passwords: Password service.\n    tokens: Token service.\n\nReturns:\n    Created public user.",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/sensors": {
      "get": {
        "description": "List sensors.\n\nArgs:\n    _: Current admin with verified TOTP.\n    store: Application store.\n\nReturns:\n    Paginated sensor list.",
        "operationId": "list_sensors_api_v1_sensors_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiPage_Sensor_"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Sensors",
        "tags": [
          "sensors"
        ]
      },
      "post": {
        "description": "Register a sensor and issue client certificate material.\n\nArgs:\n    payload: Sensor registration payload.\n    _: Current admin with verified TOTP.\n    store: Application store.\n\nReturns:\n    Registered sensor and certificate bundle.",
        "operationId": "register_sensor_api_v1_sensors_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SensorRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SensorRegisterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Sensor",
        "tags": [
          "sensors"
        ]
      }
    },
    "/api/v1/sensors/{sensor_id}": {
      "get": {
        "description": "Return one sensor.\n\nArgs:\n    sensor_id: Sensor identifier.\n    _: Current admin with verified TOTP.\n    store: Application store.\n\nReturns:\n    Sensor record.",
        "operationId": "get_sensor_api_v1_sensors__sensor_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "sensor_id",
            "required": true,
            "schema": {
              "title": "Sensor Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sensor"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Sensor",
        "tags": [
          "sensors"
        ]
      }
    },
    "/api/v1/sensors/{sensor_id}/revoke": {
      "post": {
        "description": "Revoke a sensor.\n\nArgs:\n    sensor_id: Sensor identifier.\n    _: Current admin with verified TOTP.\n    store: Application store.",
        "operationId": "revoke_sensor_api_v1_sensors__sensor_id__revoke_post",
        "parameters": [
          {
            "in": "path",
            "name": "sensor_id",
            "required": true,
            "schema": {
              "title": "Sensor Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Sensor",
        "tags": [
          "sensors"
        ]
      }
    },
...
  }
}